+++ src/main/java/com/CloudClub/jpaStudy/JpaStudyApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class JpaStudyApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(JpaStudyApplication.class, args);
-	}
+  public static void main(String[] args) {
+    SpringApplication.run(JpaStudyApplication.class, args);
+  }
 
 }
