+++ src/main/java/com/CloudClub/jpaStudy/controller/HelloController.java
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class HelloController {
 
-	@GetMapping("hello")
-	public String hello(Model model) {
-		model.addAttribute("data", "hello!");
-		return "hello";
-	}
+  @GetMapping("hello")
+  public String hello(Model model) {
+    model.addAttribute("data", "hello!");
+    return "hello";
+  }
 }
