+++ src/main/java/com/CloudClub/jpaStudy/controller/HomeController.java
@@ -1,28 +1,27 @@
 package com.CloudClub.jpaStudy.controller;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import lombok.extern.slf4j.Slf4j;
 
 @Controller
 @Slf4j
 public class HomeController {
 
-	@RequestMapping("/")
-	public String home() {
+  @RequestMapping("/")
+  public String home() {
 
-
-		log.info("home controller");
-		return "home";
-	}
+    log.info("home controller");
+    return "home";
+  }
 
   @GetMapping("/custom-login")
   public String login() {
     log.info("login controller");
     return "custom-login";
   }
+
   @GetMapping("/success")
   public String success() {
     log.info("success controller");
