--- src/main/java/com/CloudClub/jpaStudy/domain/Address.java
@@ -1,22 +1,0 @@
-package com.CloudClub.jpaStudy.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.Getter;
-
-@Embeddable
-@Getter
-public class Address {
-
-	private String city;
-	private String street;
-	private String zipcode;
-
-	protected Address() {
-	}
-
-	public Address(String city, String street, String zipcode) {
-		this.city = city;
-		this.street = street;
-		this.zipcode = zipcode;
-	}
-}
