--- src/main/java/com/CloudClub/jpaStudy/domain/Delivery.java
@@ -1,33 +1,0 @@
-package com.CloudClub.jpaStudy.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter @Setter
-public class Delivery {
-
-	@Id @GeneratedValue
-	@Column(name = "delivery_id")
-	private Long id;
-
-	@OneToOne(mappedBy = "delivery", fetch = FetchType.LAZY)
-	private Order order;
-
-	@Embedded
-	private Address address;
-
-	@Enumerated(EnumType.STRING)
-	private DeliveryStatus status;
-
-}
