--- src/main/java/com/CloudClub/jpaStudy/domain/DeliveryStatus.java
@@ -1,6 +1,0 @@
-package com.CloudClub.jpaStudy.domain;
-
-public enum DeliveryStatus {
-	READY,
-	COMP
-}
