--- src/main/java/com/CloudClub/jpaStudy/domain/OrderItem.java
@@ -1,56 +1,0 @@
-package com.CloudClub.jpaStudy.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "order_item")
-@Getter @Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OrderItem {
-
-	@Id @GeneratedValue
-	@Column(name = "order_item_id")
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "item_id")
-	private Item item;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "order_id")
-	private Order order;
-
-	private int orderPrice;
-	private int count;
-
-	// 비즈니스 로직
-	public void cancel() {
-		// 주문했던 재고를 원상복구
-		getItem().addStock(count);
-	}
-
-	public int getTotalPrice() {
-		return getOrderPrice() * count;
-	}
-
-	public static OrderItem createOrderItem(Item item, int orderPrice, int count) {
-		OrderItem orderItem = new OrderItem();
-		orderItem.setItem(item);
-		orderItem.setOrderPrice(orderPrice);
-		orderItem.setCount(count);
-
-		item.removeStock(count);
-		return orderItem;
-	}
-}
