--- src/main/java/com/CloudClub/jpaStudy/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package com.CloudClub.jpaStudy.domain;
-
-public enum OrderStatus {
-	ORDER, CANCEL
-}
