+++ src/main/java/com/CloudClub/jpaStudy/domain/delivery/Delivery.java
@@ -1,0 +1,37 @@
+package com.CloudClub.jpaStudy.domain.delivery;
+
+import com.CloudClub.jpaStudy.domain.member.Address;
+import com.CloudClub.jpaStudy.domain.member.BaseEntity;
+import com.CloudClub.jpaStudy.domain.order.Order;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToOne;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class Delivery extends BaseEntity {
+
+  @Id
+  @GeneratedValue
+  @Column(name = "delivery_id")
+  private Long id;
+
+  @OneToOne(mappedBy = "delivery", fetch = FetchType.LAZY)
+  private Order order;
+
+  @Embedded
+  private Address address;
+
+  @Enumerated(EnumType.STRING)
+  private DeliveryStatus status;
+
+}
