+++ src/main/java/com/CloudClub/jpaStudy/domain/delivery/DeliveryStatus.java
@@ -1,0 +1,6 @@
+package com.CloudClub.jpaStudy.domain.delivery;
+
+public enum DeliveryStatus {
+  READY,
+  COMP
+}
