+++ src/main/java/com/CloudClub/jpaStudy/domain/item/Album.java
@@ -1,7 +1,5 @@
 package com.CloudClub.jpaStudy.domain.item;
 
-import com.CloudClub.jpaStudy.domain.Item;
-
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Getter;
@@ -9,9 +7,10 @@ import lombok.Setter;
 
 @Entity
 @DiscriminatorValue("A")
-@Getter @Setter
+@Getter
+@Setter
 public class Album extends Item {
 
-	private String artist;
-	private String etc;
+  private String artist;
+  private String etc;
 }
