+++ src/main/java/com/CloudClub/jpaStudy/domain/item/Book.java
@@ -1,7 +1,5 @@
 package com.CloudClub.jpaStudy.domain.item;
 
-import com.CloudClub.jpaStudy.domain.Item;
-
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Getter;
@@ -9,8 +7,10 @@ import lombok.Setter;
 
 @Entity
 @DiscriminatorValue("B")
-@Getter @Setter
+@Getter
+@Setter
 public class Book extends Item {
-	private String author;
-	private String isbn;
+
+  private String author;
+  private String isbn;
 }
