+++ src/main/java/com/CloudClub/jpaStudy/domain/item/Category.java
@@ -1,0 +1,43 @@
+package com.CloudClub.jpaStudy.domain.item;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class Category {
+
+  @Id
+  @GeneratedValue
+  @Column(name = "category_id")
+  private Long id;
+
+  private String name;
+
+  @ManyToMany
+  @JoinTable(name = "category_item",
+      joinColumns = @JoinColumn(name = "category_id"),
+      inverseJoinColumns = @JoinColumn(name = "item_id"))
+  private List<Item> items = new ArrayList<>();
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "parent_id")
+  private Category parent;
+
+  @OneToMany(mappedBy = "parent")
+  private List<Category> child = new ArrayList<>();
+
+}
