+++ src/main/java/com/CloudClub/jpaStudy/domain/item/Movie.java
@@ -1,7 +1,5 @@
 package com.CloudClub.jpaStudy.domain.item;
 
-import com.CloudClub.jpaStudy.domain.Item;
-
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Getter;
@@ -9,6 +7,10 @@ import lombok.Setter;
 
 @Entity
 @DiscriminatorValue("M")
-@Getter @Setter
+@Getter
+@Setter
 public class Movie extends Item {
+
+  private String director;
+  private String actor;
 }
