+++ src/main/java/com/CloudClub/jpaStudy/domain/item/exception/NotEnoughStockException.java
@@ -1,0 +1,26 @@
+package com.CloudClub.jpaStudy.domain.item.exception;
+
+public class NotEnoughStockException extends RuntimeException {
+
+  public NotEnoughStockException() {
+    super();
+  }
+
+  public NotEnoughStockException(String message) {
+    super(message);
+  }
+
+  public NotEnoughStockException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public NotEnoughStockException(Throwable cause) {
+    super(cause);
+  }
+
+  protected NotEnoughStockException(String message, Throwable cause, boolean enableSuppression,
+      boolean writableStackTrace) {
+    super(message, cause, enableSuppression, writableStackTrace);
+  }
+
+}
