+++ src/main/java/com/CloudClub/jpaStudy/domain/order/Order.java
@@ -1,0 +1,111 @@
+package com.CloudClub.jpaStudy.domain.order;
+
+import com.CloudClub.jpaStudy.domain.delivery.Delivery;
+import com.CloudClub.jpaStudy.domain.delivery.DeliveryStatus;
+import com.CloudClub.jpaStudy.domain.member.BaseEntity;
+import com.CloudClub.jpaStudy.domain.member.Member;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "orders")
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Order extends BaseEntity {
+
+  @Id
+  @GeneratedValue
+  @Column(name = "order_id")
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "member_id")
+  private Member member;
+
+  @OneToMany(mappedBy = "order", cascade = CascadeType.ALL)
+  private List<OrderItem> orderItems = new ArrayList<>();
+
+  @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+  @JoinColumn(name = "delivery_id")
+  private Delivery delivery;
+
+  private LocalDateTime orderDate;
+
+  @Enumerated(EnumType.STRING)
+  private OrderStatus status; // 주문 상태 (ORDER, CANCEL)
+
+  // 생성 메서드
+  public static Order createOrder(Member member, Delivery delivery, OrderItem... orderItems) {
+    Order order = new Order();
+    order.setMember(member);
+    order.setDelivery(delivery);
+    for (OrderItem orderItem : orderItems) {
+      order.addOrderItem(orderItem);
+    }
+    order.setStatus(OrderStatus.ORDER);
+    order.setOrderDate(LocalDateTime.now());
+    return order;
+  }
+
+  public void setMember(Member member) {
+    this.member = member;
+    member.getOrders().add(this);
+  }
+
+  public void addOrderItem(OrderItem orderItem) {
+    orderItems.add(orderItem);
+    orderItem.setOrder(this);
+  }
+
+  public void setDelivery(Delivery delivery) {
+    this.delivery = delivery;
+    delivery.setOrder(this);
+  }
+
+  // 비즈니스 로직
+  /*
+   *
+   * 주문 취소
+   * */
+  public void cancel() {
+    if (delivery.getStatus() == DeliveryStatus.COMP) {
+      throw new IllegalStateException("이미 배송완료된 상품은 취소가 불가능합니다.");
+    }
+    this.setStatus(OrderStatus.CANCEL);
+    for (OrderItem orderItem : orderItems) {
+      orderItem.cancel();
+    }
+  }
+
+  // 조회 로직
+
+  /*
+   * 전체 주문 가격 조회
+   * */
+  public int getTotalPrice() {
+    return orderItems.stream()
+        .mapToInt(OrderItem::getTotalPrice)
+        .sum();
+  }
+
+}
+
