+++ src/main/java/com/CloudClub/jpaStudy/domain/order/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.CloudClub.jpaStudy.domain.order;
+
+public enum OrderStatus {
+  ORDER, CANCEL
+}
