--- src/main/java/com/CloudClub/jpaStudy/exception/NotEnoughStockException.java
@@ -1,25 +1,0 @@
-package com.CloudClub.jpaStudy.exception;
-
-public class NotEnoughStockException extends RuntimeException {
-	public NotEnoughStockException() {
-		super();
-	}
-
-	public NotEnoughStockException(String message) {
-		super(message);
-	}
-
-	public NotEnoughStockException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-	public NotEnoughStockException(Throwable cause) {
-		super(cause);
-	}
-
-	protected NotEnoughStockException(String message, Throwable cause, boolean enableSuppression,
-		boolean writableStackTrace) {
-		super(message, cause, enableSuppression, writableStackTrace);
-	}
-
-}
