+++ src/main/java/com/CloudClub/jpaStudy/repository/InMemoryRefreshTokenRepository.java
@@ -1,7 +1,7 @@
 package com.CloudClub.jpaStudy.repository;
 
-import com.CloudClub.jpaStudy.domain.RefreshToken;
-import com.CloudClub.jpaStudy.service.RefreshTokenRepository;
+import com.CloudClub.jpaStudy.domain.member.RefreshToken;
+import com.CloudClub.jpaStudy.repository.OrderRepository.RefreshTokenRepository;
 import java.util.HashMap;
 import java.util.Map;
 import org.springframework.stereotype.Repository;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class InMemoryRefreshTokenRepository implements RefreshTokenRepository {
 
-  private Map<String, RefreshToken> store = new HashMap<>();
+  private final Map<String, RefreshToken> store = new HashMap<>();
 
   @Override
   public void save(String username, String refreshToken, Long expiration) {
