+++ src/main/java/com/CloudClub/jpaStudy/repository/OrderRepository.java
@@ -1,14 +1,8 @@
 package com.CloudClub.jpaStudy.repository;
 
 import com.CloudClub.jpaStudy.controller.OrderSearch;
-import java.util.ArrayList;
-import java.util.List;
-
-import org.springframework.stereotype.Repository;
-import org.springframework.util.StringUtils;
-
-import com.CloudClub.jpaStudy.domain.Order;
-
+import com.CloudClub.jpaStudy.domain.member.RefreshToken;
+import com.CloudClub.jpaStudy.domain.order.Order;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.TypedQuery;
 import jakarta.persistence.criteria.CriteriaBuilder;
@@ -17,56 +11,67 @@ import jakarta.persistence.criteria.Join;
 import jakarta.persistence.criteria.JoinType;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+import org.springframework.util.StringUtils;
 
 @Repository
 @RequiredArgsConstructor
 public class OrderRepository {
 
-	private final EntityManager em;
+  private final EntityManager em;
 
-	public void save(Order order) {
-		em.persist(order);
-	}
+  public void save(Order order) {
+    em.persist(order);
+  }
 
-	public Order findOne(Long id) {
-		return em.find(Order.class, id);
-	}
+  public Order findOne(Long id) {
+    return em.find(Order.class, id);
+  }
 
-	public List<Order> findAll(OrderSearch orderSearch) {
-		// use JPQL
-		String jpql = "select o from Order o join o.member m";
+  public List<Order> findAll(OrderSearch orderSearch) {
+    // use JPQL
+    String jpql = "select o from Order o join o.member m";
 
-		return em.createQuery("select o from Order o join o.member m", Order.class)
-			.setMaxResults(1000)
-			.getResultList();
-	}
+    return em.createQuery("select o from Order o join o.member m", Order.class)
+        .setMaxResults(1000)
+        .getResultList();
+  }
 
-	/*
-	*
-	* JPA Criteria
-	*
-	* */
+  /*
+   *
+   * JPA Criteria
+   *
+   * */
 
-	public List<Order> findAllByCriteria(OrderSearch orderSearch) {
-		CriteriaBuilder cb = em.getCriteriaBuilder();
-		CriteriaQuery<Order> cq = cb.createQuery(Order.class);
-		Root<Order> o = cq.from(Order.class);
-		Join<Object, Object> m = o.join("member", JoinType.INNER);
+  public List<Order> findAllByCriteria(OrderSearch orderSearch) {
+    CriteriaBuilder cb = em.getCriteriaBuilder();
+    CriteriaQuery<Order> cq = cb.createQuery(Order.class);
+    Root<Order> o = cq.from(Order.class);
+    Join<Object, Object> m = o.join("member", JoinType.INNER);
 
-		List<Predicate> criteria = new ArrayList<>();
+    List<Predicate> criteria = new ArrayList<>();
 
-		if (orderSearch.getOrderStatus() != null) {
-			Predicate status = cb.equal(o.get("status"), orderSearch.getOrderStatus());
-			criteria.add(status);
-		}
-		if(StringUtils.hasText(orderSearch.getMemberName())) {
-			Predicate name = cb.equal(m.get("name"), "%" + orderSearch.getMemberName() + "%");
-			criteria.add(name);
-		}
-		cq.where(cb.and(criteria.toArray(new Predicate[criteria.size()])));
-		TypedQuery<Order> query = em.createQuery(cq).setMaxResults(1000);
-		return query.getResultList();
-	}
+    if (orderSearch.getOrderStatus() != null) {
+      Predicate status = cb.equal(o.get("status"), orderSearch.getOrderStatus());
+      criteria.add(status);
+    }
+    if (StringUtils.hasText(orderSearch.getMemberName())) {
+      Predicate name = cb.equal(m.get("name"), "%" + orderSearch.getMemberName() + "%");
+      criteria.add(name);
+    }
+    cq.where(cb.and(criteria.toArray(new Predicate[criteria.size()])));
+    TypedQuery<Order> query = em.createQuery(cq).setMaxResults(1000);
+    return query.getResultList();
+  }
 
+  public interface RefreshTokenRepository {
+
+    void save(String username, String refreshToken, Long expiration);
+
+    RefreshToken findByUsername(String username);
+
+  }
 }
