--- src/main/java/com/CloudClub/jpaStudy/service/ItemService.java
@@ -1,32 +1,0 @@
-package com.CloudClub.jpaStudy.service;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.CloudClub.jpaStudy.repository.ItemRepository;
-import com.CloudClub.jpaStudy.domain.Item;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class ItemService {
-
-	private final ItemRepository itemRepository;
-
-	@Transactional
-	public void saveItem(Item item) {
-		itemRepository.save(item);
-	}
-
-	public List<Item> findItems() {
-		return itemRepository.findAll();
-	}
-
-	public Item findOne(Long itemId) {
-		return itemRepository.findOne(itemId);
-	}
-}
