--- src/main/java/com/CloudClub/jpaStudy/service/MemberService.java
@@ -1,47 +1,0 @@
-package com.CloudClub.jpaStudy.service;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.CloudClub.jpaStudy.repository.MemberRepository;
-import com.CloudClub.jpaStudy.domain.Member;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class MemberService {
-
-	private final MemberRepository memberRepository;
-
-	/**
-	 * 회원 가입
-	 */
-	@Transactional
-	public Long join(Member member) {
-
-		// 중복 회원 검증
-		validateDuplicateMember(member);
-		memberRepository.save(member);
-		return member.getId();
-	}
-
-	private void validateDuplicateMember(Member member) {
-		List<Member> findMembers = memberRepository.findByName(member.getName());
-		if(!findMembers.isEmpty()) {
-			throw new IllegalStateException("이미 존재하는 회원입니다.");
-		}
-	}
-
-	//회원 전체 조회
-	public List<Member> findMembers() {
-		return memberRepository.findAll();
-	}
-
-	public Member findOne(Long id) {
-		return memberRepository.findOne(id);
-	}
-}
