+++ src/main/java/com/CloudClub/jpaStudy/service/item/ItemService.java
@@ -1,0 +1,29 @@
+package com.CloudClub.jpaStudy.service.item;
+
+import com.CloudClub.jpaStudy.domain.item.Item;
+import com.CloudClub.jpaStudy.repository.ItemRepository;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ItemService {
+
+  private final ItemRepository itemRepository;
+
+  @Transactional
+  public void saveItem(Item item) {
+    itemRepository.save(item);
+  }
+
+  public List<Item> findItems() {
+    return itemRepository.findAll();
+  }
+
+  public Item findOne(Long itemId) {
+    return itemRepository.findOne(itemId);
+  }
+}
